requirements.txt
aiohttp>=3.8.0
aiolimiter>=1.0.0
tenacity>=8.0.0
pyahocorasick>=1.4.0
aiosqlite>=0.17.0
portalocker>=2.8.0
python-dateutil>=2.8.2
pytest>=7.0.0
pytest-asyncio>=0.20.0
aioresponses>=0.7.8
fastapi>=0.95.0
uvicorn>=0.22.0
jinja2>=3.1.0
